requirements.txt
Flask
Flask-SQLAlchemy
Flask-WTF
Flask-Login
python-dotenv
email-validator
Werkzeug
yfinance